use a single MediaSourceBufferMock and MediaSourceMock definition
