+++ test/unit/mocks/MediaSourceMock.js
@@ -1,0 +1,34 @@
+import MediaSourceBufferMock from './MediaSourceBufferMock';
+
+class MediaSourceMock {
+    constructor() {
+        this.buffers = [];
+        this.readyState = 'open';
+    }
+
+    addSourceBuffer(codec) {
+        if (codec.match(/text/i)) {
+            throw new Error('not really supported');
+        }
+
+        if (codec.match(/unknown/i)) {
+            throw new Error('unknown');
+        }
+
+        let buffer = new MediaSourceBufferMock();
+        this.buffers.push(buffer);
+        return buffer;
+    }
+
+    removeSourceBuffer(buffer) {
+        let index = this.buffers.indexOf(buffer);
+
+        if (index === -1) {
+            return;
+        }
+
+        this.buffers.splice(index, 1);
+    }
+}
+
+export default MediaSourceMock;
